personalized-travel-system/backend/requirements.txt
flask
flask-cors
flask-jwt-extended
flask-sqlalchemy
mysql-connector-python
bcrypt
requests
pillow
numpy
scipy
networkx
geopy
opencv-python
scikit-learn
transformers
sentence-transformers
openai
faiss-cpu
folium
matplotlib
tqdm
pandas
flask_bcrypt
pymysql
